Split out dungeon doors
